# contract/openapi.yaml
openapi: 3.0.0
paths:
...
  /v1/roles:
    get:
      description: Requires roles:view.
      operationId: RolesController_findAll_v1
      parameters: []
      responses:
        "200":
          description: ""
      security: &a7
        - bearer: []
      summary: List roles
      tags: &a8
        - Roles
    post:
      description: Requires roles:create.
      operationId: RolesController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoleDto"
      responses:
        "201":
          description: ""
      security: *a7
      summary: Create a custom role
      tags: *a8
  /v1/roles/{id}:
    get:
      description: Requires roles:view.
      operationId: RolesController_findOne_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a7
      summary: Get a role + its granted permissions
      tags: *a8
    patch:
      description: Requires roles:edit. Built-in roles cannot be renamed.
      operationId: RolesController_update_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoleDto"
      responses:
        "200":
          description: ""
      security: *a7
      summary: Rename / edit a role
      tags: *a8
    delete:
      description: Requires roles:delete. Built-in roles → 409.
      operationId: RolesController_remove_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "204":
          description: ""
      security: *a7
      summary: Delete a custom role
      tags: *a8
  /v1/roles/{id}/permissions:
    put:
      description: Requires roles:edit.
      operationId: RolesController_setPermissions_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetRolePermissionsDto"
      responses:
        "200":
          description: ""
      security: *a7
      summary: Set a role’s module×action grants
      tags: *a8
  /v1/modules:
    get:
      description: Requires roles:view.
      operationId: RbacCatalogueController_listModules_v1
      parameters: []
      responses:
        "200":
          description: ""
      security: &a9
        - bearer: []
      summary: List system modules
      tags: &a10
        - Roles
  /v1/permissions:
    get:
      description: Requires roles:view.
      operationId: RbacCatalogueController_listPermissions_v1
      parameters: []
      responses:
        "200":
          description: ""
      security: *a9
      summary: List all (module, action) permissions
      tags: *a10
...
  /v1/pay-periods:
    get:
      description: Requires payrun:view. Pre-loaded 2026 schedule.
      operationId: PayPeriodController_list_v1
      parameters: []
      responses:
        "200":
          description: ""
      security:
        - bearer: []
      summary: List pay periods
      tags:
        - Pay Run & Holdback
  /v1/pay-runs:
    get:
      description: Requires payrun:view.
      operationId: PayRunController_list_v1
      parameters: []
      responses:
        "200":
          description: ""
      security: &a33
        - bearer: []
      summary: List pay runs
      tags: &a34
        - Pay Run & Holdback
    post:
      description: Requires payrun:create. Computes preview lines via the engine;
        nothing frozen.
      operationId: PayRunController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePayRunDto"
      responses:
        "201":
          description: ""
      security: *a33
      summary: Create / refresh a DRAFT pay run
      tags: *a34
  /v1/pay-runs/{id}:
    get:
      description: Requires payrun:view (scoped lines).
      operationId: PayRunController_findOne_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a33
      summary: Get a pay run + lines
      tags: *a34
  /v1/pay-runs/{id}/lines:
    get:
      description: Requires payrun:view (scoped).
      operationId: PayRunController_lines_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a33
      summary: Per-rep computed lines
      tags: *a34
  /v1/pay-runs/{id}/lines/{lineId}/bonus:
    post:
      description: Requires payrun:approve. Draft only.
      operationId: PayRunController_setBonus_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: lineId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetBonusDto"
      responses:
        "200":
          description: ""
      security: *a33
      summary: Set an ad-hoc bonus on a draft line
      tags: *a34
  /v1/pay-runs/{id}/finalize:
    post:
      description: "Requires payrun:approve. ATOMIC + IDEMPOTENT: freezes snapshots,
        pays sales, records/releases holdback (release timing PROPOSED — SRS
        §17), applies bonuses, composes net. Retry is a no-op."
      operationId: PayRunController_finalize_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: Idempotency-Key
          in: header
          description: Safe to retry; finalize is idempotent.
          required: false
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a33
      summary: Finalize a pay run (the money action)
      tags: *a34
  /v1/pay-runs/{id}/export:
    post:
      description: Requires payrun:export. Configurable format; marks the run
        exported; recorded via audit.
      operationId: PayRunController_export_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportPayRunDto"
      responses:
        "200":
          description: ""
      security: *a33
      summary: Generate the ADP export for a finalized run
      tags: *a34
  /v1/holdback-ledger:
    get:
      description: Requires payrun:view. Holds, schedule, release status.
      operationId: HoldbackLedgerController_list_v1
      parameters:
        - name: rep_id
          required: false
          in: query
          schema:
            type: string
        - name: status
          required: false
          in: query
          schema:
            type: string
            enum:
              - held
              - scheduled
              - released
      responses:
        "200":
          description: ""
      security:
        - bearer: []
      summary: Holdback ledger
      tags:
        - Pay Run & Holdback
...
  /v1/notifications:
    get:
      description: Authenticated; returns only the caller’s own notifications.
      operationId: NotificationsController_list_v1
      parameters:
        - name: is_read
          required: false
          in: query
          description: Filter by read/unread.
          schema:
            type: boolean
      responses:
        "200":
          description: ""
      security: &a49
        - bearer: []
      summary: My notifications
      tags: &a50
        - Reporting & Dashboards
  /v1/notifications/{id}/read:
    patch:
      description: Authenticated; only the caller’s own notification.
      operationId: NotificationsController_markRead_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a49
      summary: Mark a notification read
      tags: *a50
  /v1/notification-settings:
    get:
      description: Requires settings:view (Super Admin).
      operationId: NotificationSettingsController_list_v1
      parameters: []
      responses:
        "200":
          description: ""
      security: &a51
        - bearer: []
      summary: Notification event settings
      tags: &a52
        - Reporting & Dashboards
    patch:
      description: Requires settings:edit (Super Admin). Per-event in-app/email
        toggles; NO per-user override.
      operationId: NotificationSettingsController_update_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNotificationSettingsDto"
      responses:
        "200":
          description: ""
      security: *a51
      summary: Configure event×channel settings
      tags: *a52
...
info:
  title: Redwave ERP / HRM API
  description: Versioned REST contract for the Redwave platform. Each protected
    endpoint notes the required (module, action) permission enforced server-side
    by the RBAC guard.
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateRoleDto:
      type: object
      properties:
        name:
          type: string
          example: General Manager
        description:
          type: string
        permission_ids:
          description: Initial permission ids to grant.
          type: array
          items:
            type: string
      required:
        - name
    UpdateRoleDto:
      type: object
      properties:
        name:
          type: string
          description: Rename the role. Built-in roles cannot be renamed.
        description:
          type: string
    SetRolePermissionsDto:
      type: object
      properties:
        permission_ids:
          description: The complete set of permission ids the role should hold.
          type: array
          items:
            type: string
      required:
        - permission_ids
...
    CreatePayRunDto:
      type: object
      properties:
        pay_period_id:
          type: string
          description: The pay period to run (must be pre-loaded).
      required:
        - pay_period_id
    SetBonusDto:
      type: object
      properties:
        amount:
          type: string
          example: "50.00"
          description: Bonus amount — exact decimal STRING, never a float.
        note:
          type: string
          description: Reason / note for the bonus.
      required:
        - amount
    ExportPayRunDto:
      type: object
      properties:
        format:
          type: string
          enum:
            - csv
            - json
          default: csv
          description: Export format.
...
    EventSettingInput:
      type: object
      properties:
        event_type:
          type: string
          example: rate_change
        in_app_enabled:
          type: boolean
          example: true
        email_enabled:
          type: boolean
          example: false
      required:
        - event_type
        - in_app_enabled
        - email_enabled
    UpdateNotificationSettingsDto:
      type: object
      properties:
        settings:
          type: array
          items:
            $ref: "#/components/schemas/EventSettingInput"
      required:
        - settings
...